bot, plugin, plugins: class based callables
Behold! Plugin's decorators are now creating plugin objects, instead of
adding type unsafe attributes to functions.

This is theorically backward compatible with Sopel 8.0 and below's
system of plugin callable (i.e. functions only).

The `sopel.loader` module is still present, and still works. It should
be easier to deprecate it, to be removed in Sopel 9.

This is minor effects on the bot itself, this is more an enabler for
what's to come, opening a wide arrayd of possibilities for plugin author
in term of customization of plugin callables.

The require_* decorators are not properly handled yet, even tho they
still work "as is", they might be too dependent on the order in which
they are used to decorate a callable. This will be fixed in a future
commit.