loader, plugins: deprecate sopel.loader
The module `sopel.loader` is a relic of the past and should not be used
anymore. It is now deprecated.

Its content has been moved to `sopel.plugins.callables`.

All of its functions but `clean_module` are also deprecated and should
not be used anymore. They where useful when plugin callables and jobs
were classic functions with a twist of arbitrary attributes, something
that static typing disaproves (for good reasons).

The `clean_module` is the exception, but it will be replaced eventually
when less sheninagans are required.